refactor: delete src/data and the dead src/game/book
Two folders src/ did not need.

src/data held one file, defaultFaction.ts, behind its own @data/* path
alias — the "a folder is an addressing scheme for one file" shape #368
condemned. It is browser-only editor seed data: one production caller
(the faction create route) and six faction-editor stories/tests/
fixtures. It moves to src/app/widgets/faction-editor/defaultFaction.ts
beside those callers, the six become relative imports, the route reaches
it through @app, and the @data alias is deleted.

src/game/book was four .module.css files with zero importers — dead
since whatever removed their components. The css-orphan guard checks
classes inside imported stylesheets, not whole unimported stylesheets,
so nothing flagged them. Deleted.

Nothing else referenced @data (vite resolves aliases through
tsconfigPaths, so the tsconfig deletion covers every build). Publisher
build clean, no renderer-manifest drift.
