+++ src/app/widgets/faction-editor/defaultFaction.ts
