--- src/game/book/dreamrules/Custom.module.css
@@ -1,16 +1,0 @@
-.strongholdsMap {
-  position: relative;
-  box-sizing: border-box;
-
-  & > * {
-    display: block;
-  }
-
-  & > *:not(*:first-child) {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: auto;
-  }
-}
