--- src/game/book/utils/Book.module.css
@@ -1,14 +1,0 @@
-code {
-  -webkit-print-color-adjust: exact;
-  color-adjust: exact;
-}
-
-p {
-  -webkit-hyphens: auto;
-  -ms-hyphens: auto;
-  hyphens: auto;
-}
-
-figure {
-  break-inside: avoid;
-}
